+++ src/main/java/org/ktc2/cokaen/wouldyouin/controller/member/MemberController.java
@@ -1,6 +1,8 @@
 package org.ktc2.cokaen.wouldyouin.controller.member;
 
 import lombok.RequiredArgsConstructor;
+import org.ktc2.cokaen.wouldyouin.controller.member.dto.MemberCreateRequest;
+import org.ktc2.cokaen.wouldyouin.controller.member.dto.MemberEditRequest;
 import org.ktc2.cokaen.wouldyouin.global.ApiResponseBody;
 import org.ktc2.cokaen.wouldyouin.global.annotation.Authorize;
 import org.ktc2.cokaen.wouldyouin.service.member.MemberService;
